requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
future==0.18.2
h2o==3.30.1.3
idna==2.10
numpy==1.19.4
pandas==1.1.4
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
tabulate==0.8.7
urllib3==1.26.2
